Issue 46:  Enforce XML constraint that attribute names must be unique within an Element start/empty tag.
This checkin enforces the XML Namespace attribute uniqueness rule:
http://www.w3.org/TR/2004/REC-xml-names11-20040204/#uniqAttrs

"6.3 Uniqueness of Attributes
"In XML documents conforming to this specification, no tag may contain two attributes which:
" 1. have identical names, or
" 2. have qualified names with the same local part and with prefixes which have been bound to namespace names that are identical. 
"This constraint is equivalent to requiring that no element have two attributes with the same expanded name.
"For example, each of the bad start-tags is illegal in the following:

<!-- http://www.w3.org is bound to n1 and n2 -->
<x xmlns:n1="http://www.w3.org" 
   xmlns:n2="http://www.w3.org" >
  <bad n1:a="1"  n2:a="2" />
</x>
"
[One element dropped from their example, as it was covered in the base non-namespace XML case.]

Memory Footprint:
The Set of XML Namespace prefixes for each Element in the Stack has been generalized to a Map of prefix names to their URL strings, so that we could do the correct XML Namespace validation of expanded attribute name uniqueness.  This increases the memory footprint by a small, effectively constant, amount:  One String per level -- the maximum depth of the XML document.

Also refactored 'pendingPrefixes' into ElementMetadata, renaming it 'unverifiedNamespacePrefixes'.

Additional refactoring:  Change 'elementStack.size() == 0' to 'elementStack.empty()'.  (Left 'elementStack.size() > 0' expressions as-is, as '!elementStack.empty()' seemed harder to read.)