Issue 46:  Enforce XML constraint that attribute names must be unique within an Element start/empty tag.
Implement the 'empty namespace URL undefines the prefix' rule from sectin 6.1 of the "Namespaces in XML" REC:

http://www.w3.org/TR/2004/REC-xml-names11-20040204/#scoping

"The attribute value in a namespace declaration for a prefix MAY be empty. This has the effect, within the scope of the declaration, of removing any association of the prefix with a namespace name."

(This is the '"".equals(namespaceURL)' test in the 'getRequiredNamespaceURL' method.)

This should completely finish implementation of all functionality for this enhancement issue.

Refactoring:
 - Extracted the 'getRequiredNamespaceURL' method.
 - Deleted the unused 'isNamespacePrefixInScope' method (...which no longer had a completely correct implementation after this change anyway).